requirements.txt
pandas==2.2.3
openpyxl==3.1.5
PySimpleGUI==5.0.8.3
google-api-python-client==2.133.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
